Resolve issue #4
