--- src/main/java/Duke/Launcher.java
@@ -1,15 +1,0 @@
-//@@author lpohsien-reused
-//Resued from https://se-education.org/guides/tutorials/javaFxPart1.html
-// with minor changes.
-package Duke;
-import javafx.application.Application;
-
-/**
- * A launcher class to workaround classpath issues.
- */
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Main.class, args);
-    }
-}
-//@@author
