+++ src/main/java/duke/Launcher.java
@@ -1,0 +1,15 @@
+//@@author lpohsien-reused
+//Resued from https://se-education.org/guides/tutorials/javaFxPart1.html
+// with minor changes.
+package duke;
+import javafx.application.Application;
+
+/**
+ * A launcher class to workaround classpath issues.
+ */
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
+//@@author
