+++ src/main/java/duke/command/InvalidCommandException.java
@@ -1,0 +1,12 @@
+package duke.command;
+
+/** 
+ * This class represents the exceptions that occurs when an
+ * invalid command was given by the user.
+ */
+public class InvalidCommandException extends Exception{
+    
+    public InvalidCommandException(String errMsg) {
+        super(errMsg);
+    }
+}
