+++ src/main/java/duke/gui/MainWindow.java
@@ -1,0 +1,85 @@
+//@@author lpohsien-reused
+//Adapted from https://se-education.org/guides/tutorials/javaFxPart4.html
+// with minor changes.
+package duke.gui;
+
+import duke.Duke;
+import duke.Main;
+import javafx.fxml.FXML;
+import javafx.scene.control.Button;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+import javafx.scene.control.ScrollPane.ScrollBarPolicy;
+import javafx.scene.image.Image;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.VBox;
+/**
+ * Controller for MainWindow. Provides the layout for the other controls.
+ */
+public class MainWindow extends AnchorPane {
+    @FXML
+    private ScrollPane scrollPane;
+    @FXML
+    private VBox dialogContainer;
+    @FXML
+    private TextField userInput;
+    @FXML
+    private Button sendButton;
+
+    private Duke duke;
+
+    private Image userImage = new Image(this.getClass().getResourceAsStream("/images/user.png"));
+    private Image dukeImage = new Image(this.getClass().getResourceAsStream("/images/duke.png"));
+
+    /** The amount of time to wait after receiving the exit command from user in millisecond */
+    private final int EXIT_PAUSE_TIME = 5000;
+
+    @FXML
+    public void initialize() {
+        scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
+        scrollPane.setFitToWidth(true);
+        dialogContainer.prefWidthProperty().bind(scrollPane.widthProperty());
+        scrollPane.setOnMouseEntered(e -> {
+            scrollPane.setVbarPolicy(ScrollBarPolicy.AS_NEEDED);
+        });
+        scrollPane.setOnMouseExited(e -> {
+            scrollPane.setVbarPolicy(ScrollBarPolicy.NEVER);
+        });
+    }
+
+    public void setDuke(Duke d) {
+        duke = d;
+    }
+
+    public void showWelcome() {
+        dialogContainer.getChildren().addAll(
+            DialogBox.getDukeDialog(
+                    "Hello, welcome to duke!\n" + duke.getTaskListString(), dukeImage)
+        );
+    }
+
+    public void showResponse(String response) {
+        dialogContainer.getChildren().addAll(
+            DialogBox.getDukeDialog(response, dukeImage)
+        );
+    }
+
+    /**
+     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
+     * the dialog container. Clears the user input after processing.
+     */
+    @FXML
+    private void handleUserInput() {
+        String input = userInput.getText();
+        String response = duke.getResponse(input);
+        if (response.equals(Main.EXIT_SIGNAL)) {
+            System.exit(0);
+        }
+        dialogContainer.getChildren().addAll(
+                DialogBox.getUserDialog(input, userImage),
+                DialogBox.getDukeDialog(response, dukeImage)
+        );
+        userInput.clear();
+    }
+}
+//@@author
