+++ src/main/java/duke/storage/InvalidStorageException.java
@@ -1,0 +1,16 @@
+package duke.storage;
+
+/** 
+ * This class represents exceptions that occurs when the
+ * Storage tries to access the file.
+ */
+public class InvalidStorageException extends Exception {
+
+    /**
+     * Constructs the invalid storage exception.
+     * @param errorMsg The error message to be shown to the user.
+     */
+    public InvalidStorageException(String errorMsg) {
+        super(errorMsg);
+    }
+}
